LayoutTests/http/tests/history/redirect-303.html

<html>
<head>
<title>303 Redirect</title>
<script>
  function runTest() {
    if (window.testRunner) {
        testRunner.clearBackForwardList();
        testRunner.waitUntilDone();
    }
    window.setTimeout(function() {window.location = 'resources/redirect-helper.pl?303';}, 0);
  }
</script>

<body onload="runTest()">This page is a 303 redirect.</body>
</html>
