LayoutTests/http/tests/history/redirect-303-expected.txt

This page is the target of a redirect.

PASS: History item count should be 1 and is.
